Handle days as labels not dates
A 'day' in Garmin is a collection of data for the day, and activities on that day, so we are organising our storage and display in the same way.
This gets rid of some bugs around handling them as dates - e.g. data being displayed on the worng day etc.