Fix weird bug in first Monday in two week view
When navigating for the first time to the first Monday in the initial two week view for the first time, the two week chart would refresh.
Turned out it was because the initial two week period was being set up from current time of day, not midnight, but all other two week views were set up from midnight.